initial Lua host implementation. Configure lua scripts in the psycle.ini file, property "LuaScripts". By default it is appPath/LuaScripts. Several small build fixes and bugfixes.

+++ trunk/psycle/src/psycle/host/ChildView.cpp
@@ -1999,7 +1999,7 @@ namespace psycle { namespace host {
 			{
 				DeleteMenu(hRecentMenu, iCount, MF_BYPOSITION);
 			}
-			for(int iCount = recent.size()-1; iCount>= 0;iCount--)
+			for(int iCount = static_cast<int>(recent.size())-1; iCount>= 0;iCount--)
 			{
 				::MENUITEMINFO hNewItemInfo;
 				hNewItemInfo.cbSize		= sizeof(MENUITEMINFO);
